refactor project structure, add pdf_converter
