+++ rss_reader_files/README.md
@@ -1,0 +1,24 @@
+#Python RSS reader.
+
+#JSON structure
+<pre>
+{
+    'url': 'https://news.yahoo.com/rss/'
+    'feed': {
+        'name': 'Yahoo News - Latest News & Headlines',
+        'items': [
+            {
+                "title": "Australian police exhume body of 'spy' found dead in 1948 in hope of solving country's most enduring mystery",
+                "link": "https://news.yahoo.com/australian-police-exhume-body-spy-143207255.html",
+                "date": "21-05-19 14:32",
+                "img": [
+                    {
+                        "src": "https://s.yimg.com/ny/api/res/1.2/mqIND1K4Qx4kNCpDI5W7NQ--/YXBwaWQ9aGlnaGxhbmRlcjt3PTcwNTtoPTQ0MC41ODk3NjQwOTQzNjIyNA--/https://s.yimg.com/uu/api/res/1.2/7EFitRizm6pbIMuAg5CRxA--~B/aD0xNTYzO3c9MjUwMTthcHBpZD15dGFjaHlvbg--/https://media.zenfs.com/en/the_telegraph_258/728239dfd7d0afd0a70f711ed582055f",
+                        "alt": "The Somerton Man has inspired thousands of armchair sleuths"
+                    }
+                ]
+            } 
+        ]
+    }
+}
+</pre>
