+++ rss_reader_files/requirements.txt
@@ -1,0 +1,11 @@
+beautifulsoup4==4.9.3
+bs4==0.0.1
+certifi==2020.12.5
+chardet==4.0.0
+feedparser==6.0.2
+idna==2.10
+lxml==4.6.3
+requests==2.25.1
+sgmllib3k==1.0.0
+soupsieve==2.2.1
+urllib3==1.26.4
