+++ rss_reader_files/rss_reader.py
@@ -1,0 +1,53 @@
+import os
+import sys
+
+
+sys.path.append(os.path.dirname((os.path.dirname(__file__))))
+from rss_reader_files.argparser import get_args
+from rss_reader_files.logger_conf import create_root_logger, add_console_handler
+from rss_reader_files.RssParser import RssParser
+from rss_reader_files.cache_handlers import get_feed_from_cache, print_cached_feed
+from rss_reader_files.pdf_converter import convert_to_pdf
+
+def main():
+    arguments = get_args()
+
+    logger = create_root_logger()
+
+    if arguments.verbose:
+        add_console_handler(logger)
+
+    logger.info('Starting script')
+
+    if not arguments.url:
+        raise ValueError("URL is empty, please input URL")
+
+    logger.info(f'Program started, url: {arguments.url}')
+
+    if arguments.date:
+        feed_list = get_feed_from_cache(arguments.date, arguments.limit)
+        print_cached_feed(feed_list)
+    else:
+        get_data = RssParser(arguments.url, arguments.limit)
+        try:
+            get_data.get_feed()
+
+        except Exception as error_message:
+            logger.error(error_message)
+            if not arguments.verbose:
+                print(error_message)
+        if arguments.json:
+            print(get_data.convert_to_json())
+        elif not arguments.to_pdf:
+            print(get_data)
+        if arguments.to_pdf:
+            path_to_pdf_file = arguments.to_pdf + os.path.sep + 'news.pdf'
+            try:
+                convert_to_pdf(get_data, path_to_pdf_file)
+            except Exception as err:
+                logger.error(err)
+                if not arguments.verbose:
+                    print(err)
+
+if __name__ == '__main__':
+    main()
